Moved InactiveAwwsAggregationHelper to own file
